feat: display skipped files
Display and store skipped files as a fact, relative to the project.
The list enables us to track why some dependencies was not identified.

This is helpful, if we miss some dependency when running locally, or
if it's not identified when running a scheduled monitor.

Includes following changes:

- List first ten skipped files in display.ts, and the reason why it's skipped.
- A flag to toggle the list in display (--print-skipped-unmanaged-files)
- Sends skipped files to the registry

Ticket: TUN-96