feat: Always send half-dhash, uhash and size
To be able to better reason about which hashing technique that
works best, always hash all files with both the traditional
half-dhash and uhash. This will allow us to test both offline and
validate uhash as an alternative hashing method without source
code access.

Include size as a crude (but cheap) debugging method.