feat: Add new type for handling the new "confidence" field
The purpose of those changes are related to the addition of a new field in the
CLI output for the unmanaged flows.

When using either --print-deps or --print-dep-paths args along with the unmanaged
scan command on the CLI you should be able to see the confidence score of the
found dependencies.

Co-authored-by: daniel.ekelund@snyk.io
Co-authored-by: eleanor.kavanagh-brown@snyk.io